Extract Carousel styled components.
