+++ src/components/Carousel/Carousel.js
@@ -1,5 +1,4 @@
 import React, {useState} from 'react';
-import styled from 'styled-components';
 import ButtonSlider from 'components/ButtonSlider/ButtonSlider.style';
 
 const CarouselComponent = ({className, data}) => {
@@ -31,39 +30,39 @@ const CarouselComponent = ({className, data}) => {
     </div> );
 }
 
-const Carousel = styled(CarouselComponent)`
-    display:grid;    
-    grid-template: "track" 1fr;
-    place-items: center;
-    place-content: center;
-    overflow: hidden;
-    max-height: clamp(450px, 30vh, 600px);
+// const Carousel = styled(CarouselComponent)`
+//     display:grid;    
+//     grid-template: "track" 1fr;
+//     place-items: center;
+//     place-content: center;
+//     overflow: hidden;
+//     max-height: clamp(450px, 30vh, 600px);
 
-    & > * {
-        grid-area: track;
-        max-width: 1000px;
-    }
+//     & > * {
+//         grid-area: track;
+//         max-width: 1000px;
+//     }
 
-    .track {
-        display:grid;
-        grid-template-columns 1fr 1fr 1fr 1fr 1fr;
+//     .track {
+//         display:grid;
+//         grid-template-columns 1fr 1fr 1fr 1fr 1fr;
         
-        gap: 0.5em;
-    }
+//         gap: 0.5em;
+//     }
 
 
-    .card-container {
-        justify-self: center;
-        align-self: center;
-    }
-    img {
-        width: 100%;
-        height: auto;
-        aspect-ratio: 16 / 9;
-        object-fit: cover;
-    }
-`;
+//     .card-container {
+//         justify-self: center;
+//         align-self: center;
+//     }
+//     img {
+//         width: 100%;
+//         height: auto;
+//         aspect-ratio: 16 / 9;
+//         object-fit: cover;
+//     }
+// `;
 
 
 
-export default Carousel;
+export default CarouselComponent;
